mindAT/deepLab/requirements.txt
tensorflow-gpu==1.14.0
tensorflow-estimator==1.14.0
tensorboard==1.14.0
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.11.8
et-xmlfile==1.0.1
freeze==3.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.33.2
h5py==3.1.0
imageio==2.9.0
importlib-metadata==3.1.1
jdcal==1.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.19.3
opencv-python==4.4.0.46
openpyxl==3.0.5
Pillow==8.0.1
protobuf==3.14.0
scipy==1.5.4
six==1.15.0
termcolor==1.1.0
tifffile==2020.9.3
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
imagecodecs==2020.5.30
